Refactor: Extract HostManager from Heap God Struct
Extracted host OS resources (file I/O, sockets, processes, zip archives) from `duke_gc::Heap` into a cohesive `HostManager` in `crates/duke-gc/src/host.rs`. The `Heap` now embeds `pub host: HostManager` as a facade, separating garbage collection from host integration. Updated call-sites in `duke-interpreter` to route through `heap.host`.
